requirements.txt
# MCP Server 依賴
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
websockets==12.0
python-multipart==0.0.6

# Discord Bot 依賴
discord.py==2.3.2
python-dotenv==1.0.0

# 資料庫
sqlalchemy==2.0.23
alembic==1.13.1
sqlite3

# 工具和測試
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 日誌和監控
structlog==23.2.0
prometheus-client==0.19.0

# 安全性
cryptography==41.0.8
python-jose[cryptography]==3.3.0 